# quickstart/common/openapi.yaml
openapi: 3.0.0
info:
  title: Backend API
  description: API to manage AppInstalls and Licenses
  version: 1.0.0
paths:
...
  /app-install-requests:
    get:
      summary: List all AppInstallRequests
      operationId: listAppInstallRequests
      responses:
        '200':
          description: A list of AppInstallRequests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppInstallRequest'
        '400':
          description: Invalid input
  /app-install-requests/{contractId}/accept:
    post:
      summary: Accept an AppInstallRequest
      operationId: acceptAppInstallRequest
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
          description: The contract ID of the AppInstallRequest to accept
        - name: commandId
          in: query
          required: true
          schema:
            type: string
          description: The command ID for the ledger transaction
      requestBody:
        description: Acceptance parameters
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppInstallRequestAccept'
      responses:
        '200':
          description: AppInstall created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInstall'
        '400':
          description: Invalid input
  /app-install-requests/{contractId}/reject:
    post:
      summary: Reject an AppInstallRequest
      operationId: rejectAppInstallRequest
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
          description: The contract ID of the AppInstallRequest to reject
        - name: commandId
          in: query
          required: true
          schema:
            type: string
          description: The command ID for the ledger transaction
      requestBody:
        description: Rejection parameters
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppInstallRequestReject'
      responses:
        '200':
          description: AppInstallRequest rejected
        '400':
          description: Invalid input
  /app-install-requests/{contractId}/cancel:
    post:
      summary: Cancel an AppInstallRequest
      operationId: cancelAppInstallRequest
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: string
          description: The contract ID of the AppInstallRequest to cancel
        - name: commandId
          in: query
          required: true
          schema:
            type: string
          description: The command ID for the ledger transaction
      requestBody:
        description: Cancellation parameters
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppInstallRequestCancel'
      responses:
        '200':
          description: AppInstallRequest canceled
        '400':
          description: Invalid input
...
components:
  schemas:
...
    AppInstallRequest:
      type: object
      properties:
        contractId:
          type: string
        dso:
          type: string
        provider:
          type: string
        user:
          type: string
        meta:
          $ref: '#/components/schemas/Metadata'
    AppInstallRequestAccept:
      type: object
      required:
        - installMeta
        - meta
      properties:
        installMeta:
          $ref: '#/components/schemas/Metadata'
        meta:
          $ref: '#/components/schemas/Metadata'
    AppInstallRequestReject:
      type: object
      required:
        - meta
      properties:
        meta:
          $ref: '#/components/schemas/Metadata'
    AppInstallRequestCancel:
      type: object
      required:
        - meta
      properties:
        meta:
          $ref: '#/components/schemas/Metadata'
    AppInstall:
      type: object
      properties:
        contractId:
          type: string
        dso:
          type: string
        provider:
          type: string
        user:
          type: string
        meta:
          $ref: '#/components/schemas/Metadata'
        numLicensesCreated:
          type: integer
        licenseNum:
          type: integer
...
    Metadata:
      type: object
      properties:
        data:
          type: object
          additionalProperties:
            type: string
...